# archive/v3_cleanup_backup/cbs-platform/platform/api-schema/openapi-v2.yml
openapi: 3.0.3
info:
  title: CBS_PYTHON V2.0 - Core Banking System API
  description: |
    Advanced microservices-based Core Banking System with comprehensive financial services.
    
    ## Features
    - Microservices architecture with domain-driven design
    - OAuth2 and API key authentication
    - Real-time event streaming
    - Advanced rate limiting and throttling
    - GraphQL support for complex queries
    - Webhook notifications
    - Comprehensive audit logging
    - Multi-tenancy support
    
    ## Security
    - JWT-based authentication
    - Role-based access control (RBAC)
    - API key management
    - Request/response encryption
    - PCI DSS compliance
    
  version: 2.0.0
  contact:
    name: CBS Engineering Team
    url: https://cbs-platform.com/support
    email: api-support@cbs-platform.com
  license:
    name: Proprietary
    url: https://cbs-platform.com/license
servers:
  - url: https://api.cbs-platform.com/v2
    description: Production API
  - url: https://staging-api.cbs-platform.com/v2
    description: Staging API
  - url: http://localhost:8000/v2
    description: Development API
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - OAuth2: []
paths:
...
  /loans:
    get:
      tags:
        - Loans
      summary: List loans
      description: Get paginated list of loans
      operationId: listLoans
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by loan status
          schema:
            type: string
            enum: [draft, submitted, approved, disbursed, active, paid_off, defaulted]
        - name: type
          in: query
          description: Filter by loan type
          schema:
            type: string
            enum: [personal, home, auto, business, education]
      responses:
        '200':
          description: Loan list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanListResponse'

    post:
      tags:
        - Loans
      summary: Apply for loan
      description: Submit a loan application
      operationId: applyForLoan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationRequest'
      responses:
        '201':
          description: Loan application submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
  /loans/{loanId}:
    get:
      tags:
        - Loans
      summary: Get loan
      description: Get loan details by ID
      operationId: getLoan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LoanIdParam'
      responses:
        '200':
          description: Loan details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.cbs-platform.com/oauth/authorize
          tokenUrl: https://auth.cbs-platform.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Administrative access
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    LoanIdParam:
      name: loanId
      in: path
      required: true
      description: Loan unique identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there is a next page
        has_previous:
          type: boolean
          description: Whether there is a previous page

    # Health and System Schemas
...
    LoanApplicationRequest:
      type: object
      required:
        - customer_id
        - loan_type
        - amount
        - currency
        - term_months
        - purpose
      properties:
        customer_id:
          type: string
          format: uuid
        loan_type:
          type: string
          enum: [personal, home, auto, business, education]
        amount:
          type: number
          minimum: 1000
        currency:
          type: string
        term_months:
          type: integer
          minimum: 1
          maximum: 360
        purpose:
          type: string
          maxLength: 500
        collateral:
          type: array
          items:
            $ref: '#/components/schemas/Collateral'
        co_borrower_id:
          type: string
          format: uuid
...
    LoanResponse:
      type: object
      required:
        - id
        - customer_id
        - loan_type
        - requested_amount
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        loan_number:
          type: string
        customer_id:
          type: string
          format: uuid
        loan_type:
          type: string
          enum: [personal, home, auto, business, education]
        requested_amount:
          type: number
        approved_amount:
          type: number
        currency:
          type: string
        term_months:
          type: integer
        interest_rate:
          type: number
        status:
          type: string
          enum: [draft, submitted, approved, disbursed, active, paid_off, defaulted]
        purpose:
          type: string
        outstanding_balance:
          type: number
        next_payment_date:
          type: string
          format: date
        next_payment_amount:
          type: number
        collateral:
          type: array
          items:
            $ref: '#/components/schemas/Collateral'
        created_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
        disbursed_at:
          type: string
          format: date-time
    LoanListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LoanResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Collateral:
      type: object
      required:
        - type
        - value
        - description
      properties:
        type:
          type: string
          enum: [property, vehicle, securities, cash, other]
        value:
          type: number
        description:
          type: string
        documents:
          type: array
          items:
            type: string
            format: uri

    # Webhook Schemas
...